Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
psycopg[binary,pool]==3.2.1
azure-storage-blob==12.24.0
azure-core==1.30.2
openai==1.51.2
orjson==3.10.7